XRDDEV-1108 Move warning dialog component
