+++ src/proxy-ui-api/frontend/src/components/ui/WarningDialog.vue
