+++ src/proxy-ui-api/frontend/src/views/AddClient/FinishPage.vue
@@ -58,7 +58,7 @@
 import Vue from 'vue';
 import { mapGetters } from 'vuex';
 import LargeButton from '@/components/ui/LargeButton.vue';
-import WarningDialog from '@/views/InitialConfiguration/WarningDialog.vue';
+import WarningDialog from '@/components/ui/WarningDialog.vue';
 
 export default Vue.extend({
   components: {
